Increased test limit

+++ tests/array_generators.h
@@ -8,7 +8,7 @@ void seed_prng(uint64_t seed);
 
 
 uint64_t
-array_fill_int16_t(int32_t* a, size_t size, void (*rseed)(uint64_t),
+array_fill_int16_t(int16_t* a, size_t size, void (*rseed)(uint64_t),
                    uint64_t (*rand_uint64_t)(void));
